Move experimental code to own module
Keep code for app separate from experiments.
